+++ spring-boot-admin-docs/src/site/asciidoc/customize_http-headers.adoc
@@ -5,5 +5,5 @@ In case you need to inject custom HTTP headers into the requests made to the mon
 
 [source,java,indent=0]
 ----
-include::{samples-dir}/spring-boot-admin-sample-servlet/src/main/java/de/codecentric/boot/admin/SpringBootAdminServletApplication.java[tags=customization-http-headers-providers]
+include::{samples-dir}/spring-boot-admin-sample-servlet/src/main/java/de/codecentric/boot/admin/sample/SpringBootAdminServletApplication.java[tags=customization-http-headers-providers]
 ----
