+++ spring-boot-admin-docs/src/site/asciidoc/customize_interceptors.adoc
@@ -6,5 +6,5 @@ This can be useful for auditing or adding some extra security checks.
 
 [source,java,indent=0]
 ----
-include::{samples-dir}/spring-boot-admin-sample-servlet/src/main/java/de/codecentric/boot/admin/SpringBootAdminServletApplication.java[tags=customization-instance-exchange-filter-function]
+include::{samples-dir}/spring-boot-admin-sample-servlet/src/main/java/de/codecentric/boot/admin/sample/SpringBootAdminServletApplication.java[tags=customization-instance-exchange-filter-function]
 ----
