+++ spring-boot-admin-samples/spring-boot-admin-sample-hazelcast/src/test/java/de/codecentric/boot/admin/sample/SpringBootAdminHazelcastApplicationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package de.codecentric.boot.admin;
+package de.codecentric.boot.admin.sample;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -22,8 +22,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ExtendWith(SpringExtension.class)
-@SpringBootTest(classes = { SpringBootAdminReactiveApplication.class })
-public class SpringBootAdminReactiveApplicationTest {
+@SpringBootTest(classes = { SpringBootAdminHazelcastApplication.class })
+public class SpringBootAdminHazelcastApplicationTest {
 
 	@Test
 	public void contextLoads() {
