+++ spring-boot-admin-samples/spring-boot-admin-sample-servlet/pom.xml
@@ -101,7 +101,7 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <mainClass>de.codecentric.boot.admin.SpringBootAdminServletApplication</mainClass>
+                    <mainClass>de.codecentric.boot.admin.sample.SpringBootAdminServletApplication</mainClass>
                     <addResources>false</addResources>
                     <executable>true</executable>
                 </configuration>
