+++ spring-boot-admin-samples/spring-boot-admin-sample-servlet/src/test/java/de/codecentric/boot/admin/sample/SpringBootAdminServletApplicationTest.java
@@ -14,17 +14,19 @@
  * limitations under the License.
  */
 
-package de.codecentric.boot.admin;
+package de.codecentric.boot.admin.sample;
 
-import org.springframework.boot.actuate.endpoint.annotation.Endpoint;
-import org.springframework.boot.actuate.endpoint.annotation.ReadOperation;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-@Endpoint(id = "custom")
-public class CustomEndpoint {
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(classes = { SpringBootAdminServletApplication.class })
+public class SpringBootAdminServletApplicationTest {
 
-	@ReadOperation
-	public String invoke() {
-		return "Hello World!";
+	@Test
+	public void contextLoads() {
 	}
 
 }
